+++ tests/test_application.py
@@ -16,6 +16,7 @@ from rodi import Container, inject
 
 from blacksheep import HTTPException, JSONContent, Request, Response, TextContent
 from blacksheep.contents import FormPart
+from blacksheep.exceptions import InternalServerError
 from blacksheep.server.application import Application, ApplicationSyncEvent
 from blacksheep.server.bindings import (
     ClientInfo,
@@ -3833,11 +3834,11 @@ def test_mounting_self_raises(app):
 async def test_custom_handler_for_500_internal_server_error(app):
     @app.exception_handler(500)
     async def unhandled_exception_handler(
-        self: FakeApplication, request: Request, exc: Exception
+        self: FakeApplication, request: Request, exc: InternalServerError
     ) -> Response:
         nonlocal app
         assert self is app
-        assert isinstance(exc, TypeError)
+        assert isinstance(exc.source_error, TypeError)
         return Response(200, content=TextContent("Called"))
 
     @app.router.get("/")